Refactored constants to their own folder
